Moving IPC stuff into init
