+++ package.json
@@ -53,7 +53,7 @@
     "testRegex": "((test|spec))\\.(js|jsx)$"
   },
   "dependencies": {
-    "electron": "^1.4.1",
+    "electron": "1.8.0",
     "faker": "^4.1.0",
     "file-loader": "^0.11.2",
     "moment": "^2.18.1",
@@ -79,7 +79,7 @@
     "babel-preset-stage-0": "^6.5.0",
     "coveralls": "^2.11.14",
     "css-loader": "^0.26.1",
-    "electron-packager": "^8.5.1",
+    "electron-packager": "9.0.1",
     "enzyme": "^2.7.0",
     "enzyme-to-json": "^1.4.5",
     "eslint": "^3.10.0",
