+++ src/init.jsx
@@ -3,9 +3,10 @@ import ReactDOM from 'react-dom';
 import { applyMiddleware, createStore, combineReducers } from 'redux';
 import { Provider } from 'react-redux';
 import createLogger from 'redux-logger';
+import { ipcRenderer } from 'electron';
 
 import Application from './components/Application/Application.jsx';
-import socketInit, { socketMessageHook } from './store/socket_init.js';
+import socketInit, { socketMessageHook } from './socket.js';
 
 // Reducers
 import chat from './store/Chat';
@@ -25,12 +26,27 @@ const store = createStore(
   applyMiddleware(socketMessageHook),
 );
 
-// TODO: use a env var to disable this on build
-window.gimmeStore = store.getState;
+// TODO: Use a env var to disable this on build
+window.state = store.getState;
 
-// pass the store into a function that init's the socket biz.
-socketInit(store);
+// Configuration for the system
+ipcRenderer.on('config', (event, newConfig) => {
+  const { dispatch } = store;
+  const { config } = store.getState();
 
+  dispatch(setConfig(newConfig));
+
+  // Create Socket Connection only if there is a new apiServer config
+  if (newConfig.apiServer === config.apiServer) {
+    console.log('SOCKET', 'Already initialized, skipping ...');
+    return;
+  }
+  
+  console.log('SOCKET', 'Initializing ...');
+  socketInit(store);
+});
+
+ipcRenderer.send('init-config');
 
 ReactDOM.render(
   <Provider store={store}>
